Added event class

+++ backend/demo/src/main/java/com/example/demo/student/StudentRepo.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StudentRepo extends JpaRepository<Student,Long> {
+public interface StudentRepo extends JpaRepository<Student, Integer> {
     //@Query("SELECT s FROM Student s WHERE s.id = ? 1")
-    Optional<Student> findBySchoolID( String SchoolID);
+    Optional<Student> findBySchoolID(String SchoolID);
 
 
 }
